Se han añadido los sigueintes casos de uso:
71. Comprobar log de acciones
72. Generar informe de acceso

+++ src/main/java/es/udc/fic/ginecologia/common/security/PermissionChecker.java
@@ -4,7 +4,7 @@ import es.udc.fic.ginecologia.common.exception.InstanceNotFoundException;
 import es.udc.fic.ginecologia.model.User;
 
 public interface PermissionChecker {
-	public void checkUserExists(Long userId) throws InstanceNotFoundException;
+	public void checkUserExists(Integer userId) throws InstanceNotFoundException;
 	
 	public User checkUser(Integer id) throws InstanceNotFoundException;
